Created button in document library screen to add document with a given name
